% File man/orthodb.Rd
\name{orthodb}
\alias{orthodb}
\title{Creates an 'orthodb' class object.}
\description{Creates an 'orthodb' class object which stores the orthologous names of the genes.}
\usage{This object is created with parseData function. It has Slot1}
\arguments{
	\item{Slot1}{: 'data.frame'}
}
\author{Javier Montalvo
\email{javier.montalvo@cinvestav.mx}
}
